make python cli a package
