--- cli/sample-input-txn
@@ -1,15 +1,0 @@
-{
-  "sendPayment": {
-  "payment": {
-    "id": "r7EdsxB3N6EsJNY33X1KvQjyBvPdL3GeoFWsUaZeZYyGZcasJExpRmpeB2jdMVNynPB8rJ9ZbVXw5UNXjcxyEqNzvgVi1Pc2RMdrMqDMXKc4XCU3irAfCS8Tw2P5xPLzoJn3ecY6mPtWH3k2bw8bRu3bv7Dq29CHTKBRVx4dYxEfwecVxvgBXscuNYo5pDqgQswuazgLcfKoLGe4SW2CbQFJGyYHcHNY969yLYkMjLuVUbJrB3r3yDUN78tL9e4NM2nwn7hr2zvPPKE7iTKJ8yqtAr2rb9Z4KiUQUfzJxv9Jdx69o5UJN9NsoyU93NtwKkbdMTw46RvpsHonpww4FDz1WZXZemxPBW2LEkecGMNrLgTekXEG1hsTcPoQx8BDKbKcVmDQuRZcdcezWXRkcRSxdc69ZkSuQme6GRT7ZM7iGoVQrKPv2gAC1vZohUZMqK1KTM5TxEV3HbCAinvvATLoLoD8yobCDVULaarkTuTDU7MrZD4ZUn6v8LPfW5aWjBTGD1Av2LCMVqop77J6UCqRw8jovxqvcmDsj7DWJe6pqhfNtiCNL9w18jr3PJu6Jou8zm8QVYfYowHdXsFAkK4hyzmLGoWTZ8bqbSaPSmXCzk3jCcTLdY8SDcs5e7fVvwJh1z18nia5z7zPi6pveyCufZ3SFkJWX2TTx2Y5hUm88YAGjkCnWRyiJuDpQWmCreVLtVQFCBTwHqLif4eNHBFVKEU9QovZyJPDc",
-    "is_delegation": "False",
-    "nonce": 37,
-    "from": "tNciczxpMfZ4eW1ZPP9NVK2vxcm9cCHvTBWMe8Nskn2A25P1YqcdvFCD5LvgngraiCmAsnC8zWAiv5pwMYjrUwpMNYDePMQYiXX7HVMjrnB1JkEckyayvsAm2Bo4EQBWbHXD5Cxp65PZy5",
-    "to": "tNci9iZe1p3KK4MCcqDa52mpxBTveEm3kqZMm7vwJF9uKzGGt1pCHVNa2oMevDb1HDAs4bNdMQLNbD8N3tkCtKNGM53obE9qFkkhmqMnKRLNLiSfPJuLGsSwqnL3HxSqciJoqJJJmq5Cfb",
-    "amount": 1000,
-    "fee": 8,
-    "valid_until": 1600,
-    "memo": "2pmu64f2x97tNiDXMycnLwBSECDKbX77MTXVWVsG8hcRFsedhXDWWq"
-    }
-  }
-}
